adding an error on script if ipython folder already exists

+++ python-environ/ipython-config.sh
@@ -1,4 +1,10 @@
 #!/bin/bash
 
+if [ -e $HOME/.ipython ]
+then
+	echo "ERROR CREATE LINK IPYTHON - .ipython already exists on your home"
+	exit
+fi
+
 ln -s $PWD/dot-ipython $HOME
 mv $HOME/dot-ipython $HOME/.ipython
